Upgrading dev containter environment for development
This project use an older variant for setting up a development
environment for home assistant custom components. Which is a
non-maintailed docker image. This commit updates this to use
an approach that is very similar to what the template project
https://github.com/ludeeus/integration_blueprint sets up.

Only the minimum changes are done, the linter is left alone
as the setup seems to make sense. After starting the dev
container (as before) in vscode, an home assistant instance
with the custom component it started using either the
vscode run/debug command or running a vscode task.
If running in debug mode, then breakpoints etc should
work.

Update devcontainer to dev the integration #1