Represents xComfort physical switches with EventEntity
The xComfort physical switches does not have state, they just
create events, so they are better represented as a EventEntity.

Hass will then both make an entity with history of earlier
events, and will emit HASS events each time the switch buttons
is pushed. To get descriptive names of the entities and events,
then the each button should be named in the xComfort App.