Adds events that control another entity to the same hass-device
The physical switches in xComfort typically is mapped to lights
or shades in the xComfort app. If there is a 1-1 mapping of device
and physical switch, then this commit will find it and will add
the HASS EventEntity to the same HASS-device as the light or shade
entity use.

This makes it easier to keep track of what is connected to what in
large xComfort systems in home assistant.