Use EventEntity for xComfort physical switches (Rocker) (#67)
* Represents xComfort physical switches with EventEntity

The xComfort physical switches does not have state, they just
create events, so they are better represented as a EventEntity.

Hass will then both make an entity with history of earlier
events, and will emit HASS events each time the switch buttons
is pushed. To get descriptive names of the entities and events,
then the each button should be named in the xComfort App.

* Adds events that control another entity to the same hass-device

The physical switches in xComfort typically is mapped to lights
or shades in the xComfort app. If there is a 1-1 mapping of device
and physical switch, then this commit will find it and will add
the HASS EventEntity to the same HASS-device as the light or shade
entity use.

This makes it easier to keep track of what is connected to what in
large xComfort systems in home assistant.

* Replace the XXX models in lights and shades with better names